fixed script

+++ setup.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-git ls-files | while read file; do
+git ls-files | egrep '^\.' | while read file; do
   ln -s $PWD/$file ~
 done
 
@@ -11,4 +11,6 @@ git config --global alias.br branch
 git config --global color.ui auto
 git config --global core.excludesfile ~/.gitignore_global
 
-wget 'https://github.com/git/git/raw/master/contrib/completion/git-prompt.sh' -O $HOME/.git-prompt.sh
+if [ ! -f $HOME/.git-prompt.sh ]; then
+  wget 'https://github.com/git/git/raw/master/contrib/completion/git-prompt.sh' -O $HOME/.git-prompt.sh
+fi
